Maxima pexpect wrapper implemented
It's just a beginning, but it can do already some useful things, like
factorization, or integration and simple expressions to/from sympy.

See the docstring of the Maxima class in sympy.parsing.maxima_wrapper for
examples of usage.

Tests were implemented and they are only run if both pexpect and maxima is
present.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>